requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
frozenlist==1.6.0
gevent==25.4.2
greenlet==3.2.1
idna==3.10
multidict==6.4.3
propcache==0.3.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.1.0
twilio==9.5.2
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
